+++ scripts/reset_name.py
@@ -33,4 +33,5 @@ try:
         print("could not rename")
 
 except:
+
     print("please pass directory_name")