+++ pydub/pydub-master/CHANGELOG.md
@@ -6,7 +6,7 @@
 - Documentation improvements
 - Python 3.9 support
 - Improved efficiency of loading wave files with `pydub.AudioSegment.from_file()`
-- Ensure `pydub.AudioSegment().export()` always retuns files with a seek position at the beginning of the file
+- Ensure `pydub.AudioSegment().export()` always retuns files with a seek position at the beginning of the file   
 - Added more EQ effects to `pydub.scipy_effects` (requires scipy to be installed)
 - Fix a packaging bug where the LICENSE file was not included in the source distribution
 - Add a way to instantiate a `pydub.AudioSegment()` with a portion of an audio file via `pydub.AudioSegment().from_file()`